examples: change mouse input from dvec2 to vec2 which is sufficient

+++ examples/glcanvas.cc
@@ -59,7 +59,7 @@
 
 using namespace std::chrono;
 
-using dvec2 = glm::dvec2;
+using vec2 = glm::vec2;
 
 /* globals */
 
@@ -90,7 +90,7 @@ static bool help_text = false;
 
 struct zoom_state {
     float zoom;
-    dvec2 mouse_pos;
+    vec2 mouse_pos;
     vec2 origin;
 };
 
@@ -661,14 +661,14 @@ static void mouse_button(GLFWwindow* window, int button, int action, int mods)
 
 static void cursor_position(GLFWwindow* window, double xpos, double ypos)
 {
-    state.mouse_pos = dvec2(xpos, ypos);
+    state.mouse_pos = vec2(xpos, ypos);
 
     if (mouse_left_drag) {
         state.origin += state.mouse_pos - state_save.mouse_pos;
         state_save.mouse_pos = state.mouse_pos;
     }
     else if (mouse_right_drag) {
-        dvec2 delta = state.mouse_pos - state_save.mouse_pos;
+        vec2 delta = state.mouse_pos - state_save.mouse_pos;
         float zoom = state_save.zoom * powf(65.0f/64.0f,(float)-delta.y);
         if (zoom != state.zoom && zoom > min_zoom && zoom < max_zoom) {
             state.zoom = zoom;
